Multi-realm refactor (wip)
